TimeBlade/mcp_servers/requirements.txt
fastmcp>=0.1.0
supabase>=1.0.3
pydantic>=2.0.0
